+++ src/requests/gamedig/topHatEngineers.js
@@ -1,0 +1,7 @@
+const topHatEngineers = {
+    type: 'medievalengineers',
+    host: process.env.MEDIEVAL_DS_ADDRESS,
+    port: process.env.MEDIEVAL_DS_PORT
+}
+
+module.exports = topHatEngineers
